add `styleCleaner` to IB styles generator fix IB background opacity transitions on frontend
